+++ src/models/index.ts
@@ -26,6 +26,7 @@ export * from './interoperable-claims';
 export * from './indexing';
 export * from './identity-bootstrap';
 export * from './individual-onboarding';
+export * from './animal-onboarding';
 export * from './inter-tenant-access-contract';
 export * from './issue';
 export * from './jsonapi';